refactor: navigaiton type과 const 분리
